refactor: export accumulateEvents (from stats)
