+++ packages/cli/src/cmds/stats/accumulateEvents.ts
@@ -2,39 +2,14 @@ import * as fs from 'fs';
 import { ClassMap, Event } from '@appland/models';
 import JSONStream from 'JSONStream';
 
+const sizeof = (obj: any): number => JSON.stringify(obj).length;
+
 export type EventInfo = {
   function: string;
   count: number;
   size: number;
   location: string;
 };
-
-const KILOBYTE = 1000;
-const MEGABYTE = KILOBYTE * 1000;
-const GIGABYTE = MEGABYTE * 1000;
-
-const sizeof = (obj: any): number => JSON.stringify(obj).length;
-
-function displaySize(size: number): string {
-  let divisor: number;
-  let suffix: string;
-
-  if (size > GIGABYTE) {
-    divisor = GIGABYTE;
-    suffix = 'GB';
-  } else if (size > MEGABYTE) {
-    divisor = MEGABYTE;
-    suffix = 'MB';
-  } else if (size > KILOBYTE) {
-    divisor = KILOBYTE;
-    suffix = 'KB';
-  } else {
-    divisor = 1;
-    suffix = 'bytes';
-  }
-
-  return `${(size / divisor).toFixed(1)} ${suffix}`;
-}
 
 async function parseClassMap(mapPath: string): Promise<ClassMap> {
   let classMap: ClassMap;
@@ -55,7 +30,7 @@ async function parseClassMap(mapPath: string): Promise<ClassMap> {
   });
 }
 
-async function accumulateEvents(mapPath: string): Promise<Array<EventInfo>> {
+export async function accumulateEvents(mapPath: string): Promise<Array<EventInfo>> {
   const classMap = await parseClassMap(mapPath);
   const events = {};
 
@@ -100,27 +75,4 @@ async function accumulateEvents(mapPath: string): Promise<Array<EventInfo>> {
         })
     );
   });
-}
-
-export async function statsForMap(
-  format: string,
-  limit: number,
-  mapPath: string
-): Promise<Array<EventInfo>> {
-  const eventsStats = (await accumulateEvents(mapPath)).slice(0, limit);
-
-  if (format === 'json') {
-    console.log(JSON.stringify(eventsStats, null, 2));
-  } else {
-    eventsStats.forEach((callData, index) => {
-      const { function: fn, count, size } = callData;
-      console.log(
-        [`${index + 1}. ${fn}`, `count: ${count}`, `estimated size: ${displaySize(size)}`, ''].join(
-          '\n      '
-        )
-      );
-    });
-  }
-
-  return eventsStats;
 }
